pyproject.toml
[tool.poetry]
name = "blaise-editing-cloud-functions"
version = "0.1.0"
description = "A collection of cloud functions related to Blaise Editing Service"
authors = ["willst1 <thomas.wills@ons.gov.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
black = "^24.3.0"
pytest = "^8.1.1"
pytest-cov = "^5.0.0"
codecov = "^2.1.13"
google-cloud-logging = "^3.10.0"
mypy = "^1.10.0"
cloud-sql-python-connector = "^1.4.2"
pymysql = "^1.1.0"
sqlalchemy = "^2.0.23"
flask = "^2.0.0"


[tool.poetry.group.dev.dependencies]
mock-alchemy = "^0.2.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
